LarkFlow/requirements.txt
anthropic>=0.30.0
openai>=1.0.0
pymysql>=1.1.0
cryptography>=42.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
python-dotenv>=1.0.0
lark-oapi>=1.5.3
PyYAML>=6.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp-proto-grpc>=1.28.0